requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
asgiref==3.8.1
asyncpg==0.29.0
attrs==23.2.0
black==24.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
databases==0.9.0
distlib==0.3.8
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.4
flake8==7.0.0
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
isort==5.13.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.7.1
pydantic_core==2.18.2
pyflakes==3.2.0
PyJWT==2.8.0
pyTelegramBotAPI==4.18.1
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
telebot==0.0.5
typing_extensions==4.11.0
urllib3==2.2.1
virtualenv==20.25.1
yarl==1.9.4
